requirements.txt
discord.py
jishaku
aiofiles
flake8
